+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="2.1.0"></a>
+## 2.1.0 (2019-06-07)
+
+### Features
+
+* Add additional css classes
+* Add input status to select
+* Add .c-input__container to select
+* Add .c-input__container to textarea
+
 <a name="2.0.0"></a>
 ## 2.0.0 (2018-11-21)